fix: migration

+++ database/migrations/2022_08_15_180219_create_products_table.php
@@ -15,7 +15,7 @@ return new class () extends Migration {
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('discount_id')->nullable();
+            $table->unsignedBigInteger('discount_id')->nullable();
             $table->foreign('discount_id')->references('id')->on('discounts');
             $table->string('name');
             $table->string('handle');
